feat: render swap as a terminal screen (routes resolve through the terminal App)
The (app) layout renders <App/> and drops children, so route files under
(app)/* are stubs by design - the full swap page placed there never
mounted (/swap fell through to home). Move the quote-only UI into
terminal/screens/Swap.tsx, wire {kind:swap} through Screen +
screenForPath/pathForScreen, and keep a stub route so the URL resolves.

Swap entry lives in the live You screen fund row (account-fund button
+ onSwap prop); the YouClient edit is reverted (dead file). Settings
entry, quote API, grant marker, convertReserve and pins unchanged.
